fix: merge and rename utils files
